+++ src/i18n_check/check/unused_keys.py
@@ -117,7 +117,7 @@ def unused_keys_check(unused_keys: List[str], all_checks_enabled: bool = False) 
             "[red]\n❌ unused-keys error: There "
             + f"{to_be} {len(unused_keys)} unused i18n {key_or_keys} in the {config_i18n_src_file_name} i18n source file. Please remove or assign the following {key_or_keys}:"
             + "\n\n"
-            + "\n".join(unused_keys)
+            + "\n".join(sorted(unused_keys))
             + "[/red]"
         )
         rprint(error_message)